hassle_free_backend/requirements.txt
flask
werkzeug
transformers
torch
datasets
accelerate
pandas
spacy
nltk
python-docx
PyPDF2
scikit-learn
numpy
flask-cors
